parse: do not truncate lines at escaped #
It is sometimes useful to have # in the config option value, for ex.
when passing a hex color to a command. In this case, the user can escape
the # sign with a \, for ex.

[command]
binding = bemenu-run --fn \\#ffffff

One of the \ is consumed by wf-config, so `bemenu-run --fn \#ffffff` is
the parsed option value.