Reset options removed from config file to default
If an option is not part of the metadata, we need to reset it to default
when it is removed from the config file, or else the option will remain
stagnant in the config object, which means it exists as if it was not
removed from the config file. Fixes #26.