types: add two more edge-swipe types
On touchscreen devices, we can gain more options to trigger some actions or activate plugins by splitting the edge_swipes into 3 sections.

edge-s1-swipe: first section of the edge before the main section, for example first 20% of the edge length.
edge-swipe: main section, for example 60% of edge length centered to the edge.
edge-s2-swipe: last section of the edge after the main section, for example last 20% of the edge length.